Store Sphinx' indices and binlog in shared/
So they are kept between deploys. This is clearly documented in
https://freelancing-gods.com/thinking-sphinx/v4/deployment.html.

> Please make sure all of the above files (configuration file, pid file,
index files, binlog path) are located in a shared directory (instead of
a directory tied to a specific deployed release). Otherwise, running
rake tasks will become difficult and unreliable.

> Symlinked directories are strongly discouraged as an alternative to
(or in combination with) shared paths. Symlinked paths can be translated
to release-specific paths when generating configuration, and this can
lead to data inconsistency problems.

> You’ll want to make sure that the application shared folder has db and
tmp subdirectories (or whatever is appropriate for your settings).
You’ll also want to double check the permissions of these folders so
that the user the application and searchd both runs as can write to both
folders.

The `ts:configure` and `ts:index` are still executed at every deploy
though. I created https://github.com/coopdevs/donalo/issues/71 to keep
track of it.