Backup only a single tar.gz file
This tar file contains then the DB dump plus the app's public/ folder
with all the assets. Copying just this folder skips the codebase but
also the log/ which can be a large file and end up with a massive tar to
upload to Backblaze.

The backups_role `backups_role_assets_paths` is utterly confusing. I was
convinced it was each of the assets paths that we needed to backup (and
I believe this is how it's documented in its README) but it turns out
it's the list of paths it will pass on to `restic backup` to upload them
to the bucket.