Increase max file size
Set the Nginx's limit up to 15Mb while the app is set to 10Mb this way
requests go through and it's the app server who handles this gracefully.
Only on more desperate situations such as 15Mb we let Nginx shout with
the default error page.

Donalo complained in the first place about that page because they didn't
understand what happened after uploading a >1Mb photo. This approach
solves it.