Fix Metabase MySQL user grants
For this Ansible to append the grants we need to use `append_privs`
instead but it feels better to pass the list of grants we want, in
a declarative way, and let it handle the additions/removals.

This approach makes failures readable already, which is what we wanted.
Now we see them as:

```
fatal: [donalo.org]: FAILED! => {"changed": false, "msg": "(1146, \"Table 'donalo_production.foo' doesn't exist\")"}
```

Note we had to remove the grants for `masked_people` because the table
doesn't exist yet.